+++ src/components/ViewWindow.tsx
@@ -16,7 +16,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { lens } from "lens.ts";
 
 import ColorPicker from "./utils/ColorPicker";
-import GameViewer from "./game/GameViewer";
+import GameViewerEditor from "./game/GameViewerEditor";
 
 import { State } from "../utils/state";
 import { identity } from "../utils/utils";
@@ -46,7 +46,10 @@ export default function ViewWindow() {
   return (
     <Grid container spacing={0.2} alignItems="stretch">
       <Grid item xs={editing ? 8 : 11}>
-        <GameViewer page={pageL.format.set(format)(page)} onClick={null} />
+        <GameViewerEditor
+          page={pageL.format.set(format)(page)}
+          onClick={null}
+        />
       </Grid>
       <Grid item xs={editing ? 4 : 1}>
         <Tooltip title="edit colors" arrow>
