+++ src/components/game/StyledButton.tsx
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+
+const StyledButton = styled.button`
+  color: ${(props) => props.color};
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+`;
+
+export default StyledButton;
