+++ src/components/game/StyledImg.tsx
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+const StyledImg = styled.img`
+  max-width: 80%;
+`;
+
+export default StyledImg;
