+++ src/components/topBar/AssetsManager.tsx
@@ -18,7 +18,14 @@ import { css } from "@emotion/css";
 import { useState } from "react";
 
 import { State } from "../../utils/state";
-import { identity, openFiles, readImage, noExt } from "../../utils/utils";
+import {
+  identity,
+  openFiles,
+  readImage,
+  noExt,
+  assetPath,
+  getExtensions,
+} from "../../utils/utils";
 import ImageList from "@mui/material/ImageList";
 import ImageListItem from "@mui/material/ImageListItem";
 
@@ -30,24 +37,6 @@ const StyledMiniature = styled.img`
   max-width: 80px;
   max-height: 40px;
 `;
-
-const assetPath = (assetType: string, assetName: string) =>
-  `assets/${assetType}/${assetName}`;
-
-const getExtensions = (assetType: string) => {
-  switch (assetType) {
-    case "image":
-      return [
-        "image/jpeg",
-        "image/gif",
-        "image/bmp",
-        "image/png",
-        "image/webp",
-      ];
-    default:
-      return [];
-  }
-};
 
 interface CompProps {
   open: boolean;
